+++ src/Server/Server.cpp
@@ -6,81 +6,6 @@ void Server::disconnectSocket(int socket) { close(socket); }
 
 const std::vector<t_multi_server> &Server::get_servers(void) { return servers; }
 
-void Server::addCloseFdVector(int fd)
-{
-  if (m_close_udata_map.find(fd) == m_close_udata_map.end())
-  {
-    std::vector<t_event_udata*> vec;
-    m_close_udata_map.insert(std::make_pair(fd , vec));
-    std::cout << "make new vector fd :" << fd << std::endl;
-  }
-}
-
-void Server::addUdataContent(int fd, t_event_udata *udata)
-{
-  std::map<int, std::vector<t_event_udata *> >::iterator it;
-  
-  if (m_close_udata_map.find(fd) == m_close_udata_map.end())
-  {
-    std::vector<t_event_udata*> vec;
-    m_close_udata_map.insert(std::make_pair(fd , vec));
-    std::cout << "make new vector fd :" << fd << std::endl;
-  }
-  it = m_close_udata_map.find(fd);
-  if (it != m_close_udata_map.end())
-  {
-    it->second.push_back(udata);
-  }
-}
-
-void Server::clearUdataContent(int fd, t_event_udata *udata)
-{
-  std::cout << "claerUdataContent" << std::endl;
-  std::map<int, std::vector<t_event_udata *> >::iterator it;
-  it = m_close_udata_map.find(fd);
-  if (it != m_close_udata_map.end())
-  {
-    for (size_t i = 0; i < it->second.size(); ++i)
-    {
-      printf("%p %p \n",it->second[i], udata);
-      if (it->second[i] == udata){
-        std::cout << "in here" << std::endl;
-         it->second[i] = NULL;
-      }
-    }
-  }
-}
-
-void Server::clearUdata()
-{
-  std::cout << "claerUdata" << std::endl;
-  for (std::set<int>::iterator i = m_close_fd_vec.begin(); i != m_close_fd_vec.end(); ++i)
-  {
-    std::map<int, std::vector<t_event_udata *> >::iterator it;
-    it = m_close_udata_map.find(*i);
-    if (it == m_close_udata_map.end())
-    {
-      continue;
-    }
-
-    for (size_t j = 0; j < it->second.size(); ++j)
-    {
-      if (it->second[j] != NULL)
-      {
-        printf("delete %p \n", it->second[j]);
-        delete it->second[j];
-        addEventToChangeList(m_kqueue.change_list, *i, EVFILT_WRITE,
-                       EV_DELETE, 0, 0, NULL);
-      }
-    }
-    it->second.clear();
-    m_close_udata_map.erase(it);
-    disconnectSocket(*i);
-  } 
-  m_close_fd_vec.clear();
-  std::cout << "end" << std::endl;
-}
-
 void Server::addEventToChangeList(
     std::vector<struct kevent> &change_list,
     uintptr_t ident, /* identifier for this event */
@@ -101,68 +26,35 @@ e_kqueue_event getEventStatus(struct kevent *current_event, e_event_type type)
   if (current_event->flags & EV_ERROR)
   {
     if (type == SERVER)
-    {
-      LOG_DEBUG("SERVER ERROR");
       return SERVER_ERROR;
-    }
     else if (type == CLIENT)
-    {
-      LOG_DEBUG("CLIENT ERROR");
       return CLIENT_ERROR;
-    }
   }
   if (current_event->filter == EVFILT_TIMER)
   {
-    {
-      LOG_DEBUG("CGI TIMEOUT");
-      return CGI_PROCESS_TIMEOUT;
-    }
+    return CGI_PROCESS_TIMEOUT;
   }
   if (current_event->filter == EVFILT_READ)
   {
     if (type == SERVER)
-    {
-      LOG_DEBUG("SERVER READ");
       return SERVER_READ;
-    }
     else if (type == CLIENT)
-    {
-      LOG_DEBUG("CLIENT READ");
       return CLIENT_READ;
-    }
     else if (type == PIPE)
-    {
-      LOG_DEBUG("PIPE READ");
       return PIPE_READ;
-    }
     else if (type == STATIC_FILE)
-    {
-      LOG_DEBUG("STATIC FILE READ");
       return STATIC_FILE_READ;
-    }
   }
   if (current_event->filter == EVFILT_WRITE)
   {
     if (type == SERVER)
-    {
-      LOG_DEBUG("SERVER WRITE");
       return SERVER_WRITE;
-    }  
     else if (type == CLIENT)
-    {
-      LOG_DEBUG("CLIENT WRITE");
       return CLIENT_WRITE;
-    }  
     else if (type == PIPE)
-    {
-      LOG_DEBUG("PIPE WRITE");
       return PIPE_WRITE;
-    }  
     else if (type == STATIC_FILE)
-    {
-      LOG_DEBUG("STATIC FILE WRITE");
       return STATIC_FILE_WRITE;
-    }  
   }
   return NOTHING;
 }
@@ -177,9 +69,9 @@ void Server::addServerSocketEvent(std::vector<t_multi_server> &servers,
     {
       udata = new t_event_udata(SERVER, server_conf.get_m_server_conf());
     }
-    catch(std::exception e)
+    catch (std::exception e)
     {
-        std::cout << e.what() << std::endl;
+      std::cout << e.what() << std::endl;
       exit(EXIT_FAILURE);
     }
     addEventToChangeList(m_kqueue.change_list, servers[i].server_sock,
@@ -191,9 +83,7 @@ Server::Server(Config &server_conf)
 {
   setServers(server_conf, servers);
   setSocket(servers);
-
   startBind(servers);
-
   startListen(servers, BACK_LOG);
 
   m_kqueue.kq = getKqueue();
@@ -203,12 +93,9 @@ Server::Server(Config &server_conf)
 
 Server::Server() {}
 
-Server::Server(const Server &other)
-{
-  *this = other;
-}
+Server::Server(const Server &other) { *this = other; }
 
-Server::~Server() { }
+Server::~Server() {}
 
 Server &Server::operator=(const Server &other)
 {
@@ -233,12 +120,15 @@ void Server::start(void)
     current_events = kevent(m_kqueue.kq, &m_kqueue.change_list[0],
                             m_kqueue.change_list.size(), m_kqueue.event_list,
                             MAX_EVENT_LIST_SIZE, NULL);
+
     if (current_events == -1)
     {
-      ft_error_exit(1, strerror(errno));
+      ft_error_exit(EXIT_FAILURE, strerror(errno));
     }
+    
     m_kqueue.change_list.clear();
-    std::cout << "cycle start --------------------------" << std::endl << std::endl;
+    std::cout << "cycle start --------------------------" << std::endl
+              << std::endl;
     for (int i = 0; i < current_events; ++i)
     {
       current_event = &m_kqueue.event_list[i];
@@ -246,16 +136,16 @@ void Server::start(void)
 
       if (current_event->flags & EV_ERROR)
       {
-        printf("filter: %d ident: %d \n", current_event->filter, current_event->ident);
-        std::cout << "flag is EV_ERROR: " << current_event->ident << " " << strerror(current_event->data) << std::endl;
-        printf("ERROR POINTER IS %p\n", current_event->udata);
+        printf("filter: %d ident: %d \n", current_event->filter,
+               current_event->ident);
+        // std::cout << "flag is EV_ERROR: " << current_event->ident << " " <<
+        // strerror(current_event->data) << std::endl; printf("ERROR POINTER IS
+        // %p\n", current_event->udata);
         continue;
       }
-      else
-      {
-        printf("filter: %d ident: %d  type:%d \n", current_event->filter, current_event->ident, event_status);
-        event_status = getEventStatus(current_event, current_udata->m_type);
-      }
+      event_status = getEventStatus(current_event, current_udata->m_type);
+      printf("filter: %d ident: %d  type:%d \n", current_event->filter,
+              current_event->ident, event_status);
       switch (event_status)
       {
         case SERVER_READ:
@@ -273,6 +163,18 @@ void Server::start(void)
         case CLIENT_READ:
         {
           clientReadEvent(current_event);
+          break;
+        }
+
+        case CLIENT_WRITE:
+        {
+          clientWriteEvent(current_event);
+          break;
+        }
+
+        case CLIENT_ERROR:
+        {
+          disconnectSocket(current_event->ident);
           break;
         }
 
@@ -306,27 +208,17 @@ void Server::start(void)
           break;
         }
 
-        case CLIENT_WRITE:
-        {
-          clientWriteEvent(current_event);
-          break;
-        }
-
-        case CLIENT_ERROR:
-        {
-          disconnectSocket(current_event->ident);
-          break;
-        }
-
         default:
         {
           break;
         }
       }
     }
-    std::cout << std::endl << "-------------------------- cycle END " << std::endl << std::endl;
+    std::cout << std::endl
+              << "-------------------------- cycle END " << std::endl
+              << std::endl;
 
-    if (m_close_fd_vec.size() > 0)
+    if (m_close_fd_set.size() > 0)
     {
       clearUdata();
     }
